Notification generation in friend request and accept is implemented

+++ profile.php
@@ -23,7 +23,8 @@
       if(isset($_POST['add_friend'])) {
         $my_id = $_SESSION["user_id"];
           $sql = "INSERT INTO friend_requests (fr_receiver_id, fr_sender_id) VALUES ('$id', '$my_id')";
-          if(mysqli_query($con, $sql)) {
+          $sql2 = "INSERT INTO notifications (notification_type, notification_receiver_id, notification_sender_id, notification_message) VALUES ('friend_request', '$id', '$my_id', 'sent you a friend request')";
+          if(mysqli_query($con, $sql) && mysqli_query($con, $sql2)) {
             header("Location: profile.php?id=$id");
           }
       }
@@ -51,7 +52,8 @@
       if(isset($_POST['accept_friend_request'])) {
         $my_id = $_SESSION["user_id"];
         $sql = "INSERT INTO friends (fr_receiver_id, fr_sender_id) VALUES ('$my_id', '$id')";
-        if(mysqli_query($con, $sql)) {
+        $sql2 = "INSERT INTO notifications (notification_type, notification_receiver_id, notification_sender_id, notification_message) VALUES ('friend_request_accepted', '$id', '$my_id', 'accepted your friend request')";
+        if(mysqli_query($con, $sql) && mysqli_query($con, $sql2)) {
           $sql = "DELETE FROM friend_requests WHERE fr_receiver_id='$my_id' AND fr_sender_id='$id'";
           mysqli_query($con, $sql);
           header("Location: profile.php?id=$id");
